# docs/contracts/pilot.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QatuPOS staff laboratory pilot API",
    "version": "0.2.0",
    "description": "Single branch cloud-writer laboratory, PostgreSQL real state. Not G2, no real payments, printing, SUNAT issuance, Qatu/Delivery/QR or offline claims. UUID IDs and integer PEN cents, whole sale quantities only. Client cannot supply tenant/branch/role/price; scope from session. Commands strict; same actor+scope+operation_id+canonical body returns stable result without duplicate effects, different body conflicts. Local guest key increment: only waiter activates; positive fully-paid check settles access; browser-private guest orders reuse core writer. QR/NFC hardware, public LAN and real providers remain disabled. Corrective011: quotes single-use and actor-bound, expiry exclusive. Simulated fiscal documents do not establish fiscal acceptance: check remains pending. NC supports total01/06 and02factura only;03/07 return UNSUPPORTED_CAPABILITY. No fake QR, XML signature, transmission or certification."
  },
  "security": [
    {
      "staffSession": []
    }
  ],
  "paths": {
    "/v1/pos/session": {
      "post": {
        "operationId": "loginStaff",
        "summary": "Personal password login in laboratory",
        "security": [],
        "description": "Server membership resolves role/scope. Rate-limit failures, compare password hashes, require same-site Origin, rotate cookie; no role picker grants access. Do not log password. Synthetic seed accounts only.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Established session; Set-Cookie Secure in HTTPS, HttpOnly SameSite=Strict Path=/; session token never in response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Strict invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          },
          "401": {
            "description": "Credentials rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          },
          "429": {
            "description": "Too many failed logins",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "currentStaffSession",
        "responses": {
          "200": {
            "description": "Session identity and CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "logoutStaff",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 32
            },
            "description": "Bound to authenticated session. Origin checked; no token in URLs."
          }
        ],
        "description": "Revoke persisted session and expire cookie. Login/logout do not alter financial operations.",
        "responses": {
          "204": {
            "description": "Revoked session"
          },
          "401": {
            "description": "Unauthenticated session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          },
          "403": {
            "description": "Role, scope, owner, station or CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pos/snapshot": {
      "get": {
        "operationId": "getPosSnapshot",
        "description": "Session-scoped snapshot, Cache-Control no-store. Never transport hidden financial records to waiter/kitchen. Counting participants get null expected cash/on-hand/reserved/available until declaration sealed; raw state remains numeric. Admin has same blind redaction during own participation. Kitchen sees only assigned station order lines/jobs; waiter sees table check charge/paid/held summary but no payment evidence, cash, custody, audits. staff metadata minimal no passwords. No tenant/role query override.",
        "responses": {
          "200": {
            "description": "Role-filtered operational snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          },
          "403": {
            "description": "Role, scope, owner, station or CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/guest/session": {
      "post": {
        "operationId": "post__v1_guest_session",
        "security": [],
        "responses": {
          "200": {
            "description": "Local laboratory result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Strict request rejected"
          },
          "401": {
            "description": "Invalid key/session"
          },
          "403": {
            "description": "Forbidden or guest access concluded"
          },
          "409": {
            "description": "Concurrent update / idempotency conflict"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestCodeRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "get__v1_guest_session",
        "security": [
          {
            "guestCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Local laboratory result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Strict request rejected"
          },
          "401": {
            "description": "Invalid key/session"
          },
          "403": {
            "description": "Forbidden or guest access concluded"
          },
          "409": {
            "description": "Concurrent update / idempotency conflict"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete__v1_guest_session",
        "security": [
          {
            "guestCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "Session ended"
          },
          "400": {
            "description": "Strict request rejected"
          },
          "401": {
            "description": "Invalid key/session"
          },
          "403": {
            "description": "Forbidden or guest access concluded"
          },
          "409": {
            "description": "Concurrent update / idempotency conflict"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/guest/snapshot": {
      "get": {
        "operationId": "get__v1_guest_snapshot",
        "security": [
          {
            "guestCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Local laboratory result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Strict request rejected"
          },
          "401": {
            "description": "Invalid key/session"
          },
          "403": {
            "description": "Forbidden or guest access concluded"
          },
          "409": {
            "description": "Concurrent update / idempotency conflict"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/guest/orders": {
      "post": {
        "operationId": "post__v1_guest_orders",
        "security": [
          {
            "guestCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Local laboratory result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Strict request rejected"
          },
          "401": {
            "description": "Invalid key/session"
          },
          "403": {
            "description": "Forbidden or guest access concluded"
          },
          "409": {
            "description": "Concurrent update / idempotency conflict"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestOrderInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/pos/guest-access/{visit_id}": {
      "get": {
        "operationId": "get__v1_pos_guest_access_visit_id",
        "security": [
          {
            "staffSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Local laboratory result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestAccessView"
                }
              }
            }
          },
          "400": {
            "description": "Strict request rejected"
          },
          "401": {
            "description": "Invalid key/session"
          },
          "403": {
            "description": "Forbidden or guest access concluded"
          },
          "409": {
            "description": "Concurrent update / idempotency conflict"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/v1/pos/quotes": {
      "post": {
        "operationId": "createStaffQuote",
        "security": [
          {
            "staffSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-calculated official quote with 120s TTL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Strict request rejected"
          },
          "401": {
            "description": "Invalid session"
          },
          "403": {
            "description": "Forbidden or visit closed"
          },
          "404": {
            "description": "Product or visit not found"
          },
          "409": {
            "description": "Product unavailable"
          },
          "503": {
            "description": "Base indisponible o copia restaurada en cuarentena; no se admiten operaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Required with exact Origin in operational mode; validated against the authenticated staff/guest session."
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "staffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qatu_session"
      },
      "guestCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qatu_guest"
      }
    },
    "schemas": {
...
      "SessionResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/StaffUser"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "csrf_token": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          }
        },
        "required": [
          "user",
          "expires_at",
          "csrf_token"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "GuestCodeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "minLength": 10,
            "maxLength": 32
          }
        },
        "required": [
          "code"
        ]
      },
...
    }
  },
...
}